==> 05_sentiment.ipynb <==
# Translate FR to EN

import numpy as np
import pandas as pd
print(np.__version__)
print(pd.__version__)

from transformers import MarianMTModel, MarianTokenizer
src_text = [
    "bonjour je voudrais des chips s'il vous plaît",
    "serait-il possible de faire moins de bruit ?",
    "les burgers sont-ils cuits ?",
    "et ça vous fait rire !"
]
src_text

model_name = 'Helsinki-NLP/opus-mt-fr-en'
marianmt_tokenizer = MarianTokenizer.from_pretrained(model_name)

marianmt_model = MarianMTModel.from_pretrained(model_name)
translated = marianmt_model.generate(**marianmt_tokenizer(src_text, return_tensors="pt", padding=True))

translation = [marianmt_tokenizer.decode(t, skip_special_tokens=True) for t in translated]

translation

# Predict sentiment (for twitter)

from transformers import AutoModelForSequenceClassification
from transformers import TFAutoModelForSequenceClassification
from transformers import AutoTokenizer
import numpy as np
from scipy.special import softmax
import csv
import urllib.request

# Preprocess text (username and link placeholders)
def preprocess(text):
    new_text = []
 
 
    for t in text.split(" "):
        t = '@user' if t.startswith('@') and len(t) > 1 else t
        t = 'http' if t.startswith('http') else t
        new_text.append(t)
    return " ".join(new_text)

# Tasks:
# emoji, emotion, hate, irony, offensive, sentiment
# stance/abortion, stance/atheism, stance/climate, stance/feminist, stance/hillary

task='sentiment'
MODEL = f"cardiffnlp/twitter-roberta-base-{task}"

sentiment_tokenizer = AutoTokenizer.from_pretrained(MODEL)
sentiment_tokenizer.save_pretrained(MODEL)

# download label mapping
labels=[]
mapping_link = f"https://raw.githubusercontent.com/cardiffnlp/tweeteval/main/datasets/{task}/mapping.txt"
with urllib.request.urlopen(mapping_link) as f:
    html = f.read().decode('utf-8').split("\n")
    csvreader = csv.reader(html, delimiter='\t')
labels = [row[1] for row in csvreader if len(row) > 1]

# PT
sentiment_model = AutoModelForSequenceClassification.from_pretrained(MODEL)
sentiment_model.save_pretrained(MODEL)



def preprocess_and_predict_sentiment(txt):
    txt = preprocess(txt)
    encoded_input = sentiment_tokenizer(txt, return_tensors='pt')
    output = sentiment_model(**encoded_input)
    scores = output[0][0].detach().numpy()
    scores = softmax(scores)
    return scores

text

text = "Good night 😊"

preprocess_and_predict_sentiment(text).shape


[preprocess_and_predict_sentiment(text) for text in translation]

ranking = np.argsort(scores)[::-1]
for i in range(scores.shape[0]):
    l = labels[ranking[i]]
    s = scores[ranking[i]]
    print(f"{i+1}) {l} {np.round(float(s), 4)}")
