==> 06_classification_finetuning.ipynb <==
from transformers import AutoTokenizer, AutoModel

tokenizer = AutoTokenizer.from_pretrained("camembert-base")
model = AutoModel.from_pretrained("camembert-base")

inputs = tokenizer("J'aime le fromage!", return_tensors="pt")
outputs = model(**inputs)
